issue_ui.rb: add hackish fix for odd DateTime parsing
If you use DateTime.parse, and you provide a date without a time, and convert
it to a Time object using to_time, you get a weird date, set back by 5 hours
(I suspect this is UTC vs local time).

This is a quick hacky fix that adds 9:00 am at the end of the string in order
to force the correct date. I should kinda fix a better version of this edit
in the future, but for now this will do.