log_entry: remove turntables dead code
Formerly, wlog was using turntables in order to create the databases.
This meant that whatever database operation had to be done, we would
have to write our own SQL, and logic to convert retrieved rows into
objects in memory.

There was a pretty big migration in v1.1.5 where this library was
ripped out and replaced with activerecord if you are curious to see
what all of this was about.